correct the open web platform wikipedia link
This change fixes the missing link for the "Open Web Platform" reference
in the notes from the "What is a Frontend Developer" section. An id
based link is created for the wikipedia page to ensure that both links
on the page refer to the same URL.

It also reformats the horizontal rule above the section to fix a syntax
bug which appears to make the notes section render as a single link.